requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
click==8.0.1
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
dj-database-url==0.5.0
Django==3.2.5
django-cors-headers==3.7.0
django-heroku==0.3.1
django-jwt-auth==0.0.2
django-render-block==0.8.1
django-shortuuidfield==0.1.3
django-templated-email==3.0.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.2
future==0.18.2
gunicorn==20.1.0
kombu==5.1.0
prompt-toolkit==3.0.19
psycopg2==2.9.1
PyJWT==2.1.0
pytz==2021.1
redis==3.5.3
serializers==0.2.4
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
drf-yasg==1.20.0
Pillow==8.3.1
python-dateutil==2.8.1